ENH: Add the idea of iterator axis to SumMultiply
The idea is that one plate axis could be used as an iterator. This
allows SumMultiply to compute the result in smaller blocks instead of a
possibly huge resulting array. This iterator could be used, for
instance, in GaussianMarkovChain for the state dynamics matrix if it
is different at each time instance. Because the state dynamics matrix is
needed only for one time instance at a time.