ENH: Implement plate rotation speed-up for Gaussian array node.
Rotation transformations can speed up the inference of linear
state-space models significantly. The rotation requires the rotation of
the state dynamics matrix which leads to also rotating the plate axes in
addition to the variable dimension axes. This plate rotation is now
implemented in GaussianArrayARD for performing the rotation and
RotateGaussianArrayARD for optimizing the rotation. Previous
implementation in RotateGaussianARD is deprecated because the new
implementation is more general.

See: Fast Variational Bayesian Linear State-Space Model (Luttinen, ECML
2013)