BUG: Fix speed-up rotations for Gaussian ARD arrays.
Rotation transformations for speeding up VB learning of Gaussian ARD
arrays contained several bugs when using the transformations for
variables with multiple variable axes or plate axes. In addition, alpha
hyperparameter was not handled properly unless it had plates only for
the rotated variable axis. This collection of bugs should now be fixed
and RotateGaussianArrayARD can be used for large range of different
model blocks.