DEMO: Add command line options to LSSM demo
Linear state-space model (LSSM) demo (bayespy/demos/demo_lssm.py) can be
run as
    python demo_lssm.py <options>
with several command line arguments, such as dimensionality of the data
vectors or the latent space of the model and whether to use speed-up
rotations or not. It is highly recommended to use rotations, otherwise
it may take a very long time to reach convergence.