ENH: Improve computations for plate rotation of Gaussian ARD arrays.
In the rotation speed-up of Gaussian ARD arrays, some moments can be
computed only once per optimization/iteration. This has now been
utilized better for plate rotation. It is possible to choose whether to
compute those moments at the beginning of the optimization (precompute
flag) or at each iteration step. The computations are performed a bit
differently depending on which method one chooses, but usually it is
better not to pre-compute but instead compute them once during each
iteration. If the rotation dimensionalities for both Q and R are small,
and X has other large axes that are not in alpha, then precomputation
might be useful.