BUG: Slice node modified the moment list of the parent node
Slice operator node accidentally modified the moment list of its parent
when computing the moments. This is now fixed by sending a copy of the
moment list from stochastic nodes and by taking a copy of the moment
list of the parents when receiving messages from them. Also, a test is
added which makes sure that this works for the Slice node.