ENH: Add plotting utility to nodes
When creating a node, you can give a function as a keyword argument
`plotter`. This function plots the node. So, creating a node as:
  X = Node(plotter=my_plotter_func)
and using the new plot method as:
  X.plot(kwarg1='foo', kwarg2='bar')
is identical to calling
  my_plotter_func(X, kwarg1='foo', kwargs2='bar')
Thus, it is easy to plot the distribution of the node by just calling
`plot` method.