ENH: Add plotting and monitoring to VB inference class
Class VB has now method `plot` which can be used to plot the
distribution of the nodes in the model. In order to plot a node, it must
have a plotter defined. You can also plot only some nodes by giving the
nodes or their names as arguments. In addition, VB.update plots the
distribution of a node after updating if `plot=True` is given. Because
a node is plotted to a same figure each time, this enables easy and
visual way to monitor the VB algorithm.