ENH: Add classes which define the sufficient statistics
These classes define the messages to children. Separating this protocol
from the nodes gains a few benefits:
  * Write statistic-specific features in one place only. For instance,
    covariance from Gaussian message.
  * Different nodes may have identically defined statistic so you need
    to implement related features only once. For instance, Gaussian and
    GaussianARD differ on the prior but the moments are the same.
  * General processing nodes which do not change the type of the moments
    may "inherit" the features from the parent node. For instance, slicing operator.
  * Conversions can be done easily in both of the above cases if the
    message conversion is defined in the moments class. For instance,
    GaussianMarkovChain to Gaussian and GaussianDriftingMarkovChain to
    Gaussian.

The base class will handle the necessary relevant tasks now so it is not
necessary to define the class for nodes. However, in the future some
calculations might be moved from nodes to statistics classes.

Currently, only GaussianStatistics and GaussianMarkovChainStatistics are
implemented.