BUG: Fix sharex handling in timeseries plotting
Previously, if plotting multiple sets of timeseries, the axes were
cleared before each plot. Thus only one set of timeseries was shown, the
last one. This was because, if subplot gets a different value for sharex
than before, it will create a new axes object. Now, the plotting is done
so that the sharex value remains the same and the user can thus plot
multiple sets of timeseries in one figure.