ENH: Add delete method to nodes.
The method makes it possible to delete a whole branch from a model, that
is, a node and all its child nodes recursively. Maybe not very useful
for the end user, but the main usage could be the following: When
converting nodes to proper types just so that it is simpler to use, the
conversion node does not disappear but remains as a child of the
original node. Thus, there can be several "haunting" unused conversion
nodes. So, if the conversion node is not really used anymore, delete it.