MAINT: Rename Statistics to Moments
The expectation of the sufficient statistic vector is a common term in
VB algorithms. Previously, a class for defining those was called
Statistics and its sub-classes, e.g., GaussianStatistics. However, in
Winn&Bishop paper about VMP, they use the term 'moments'. Thus,
Statistics is now called Moments and GaussianStatistics is called
GaussianMoments and so on.