API: Change GaussianARD API
GaussianARD now accepts GaussianGammaARD parents. Also, the shape of the
mean parent is interpreted as a concatenation of the plates and the
dims. This is simpler than before and allows one to use fully factorized
mean for variables that are not fully factorized (this works because the
ARD prior is fully factorized).