DOC: Make doc compiler compatible with libre systems
Sphinx is used to compile the documentation to various formats.
When compiling to PDF using LaTeX, Sphinx uses the package fontenc
(T1) by default. This does not work on fully free operating systems
such as Parabola or Trisquel. Thus, this package was removed
because the compiling should work without it.