API: Simplify GaussianARD mean parent handling
Previously, the shape/ndim of GaussianARD was inferred automatically
based on the shape/ndim of the parent mean. This was unnecessary because
those shapes/ndims have no connection whatsoever. It restricted creating
GaussianARD nodes that have smaller ndim than their parent mean although
there is no reason for that.

Now, the mean parent is mapped to fully factorized Gaussian moments and
the default is to use fully factorized moments for the node itself. The
user needs to explicitly provide shape or ndim to have different
factorization.